split up some files
